+++ apps/testingaccessibility/src/components/pricing.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react'
-import {type FormattedPrice} from '@skillrecordings/commerce-server'
 import {usePriceCheck} from '../context/pricing-check-context'
-import type {SanityProduct} from '@skillrecordings/commerce-server'
+import type {
+  SanityProduct,
+  FormattedPrice,
+} from '@skillrecordings/commerce-server/dist/@types'
 import {CheckCircleIcon} from '@heroicons/react/outline'
 import {getCouponLabel} from 'utils/get-coupon-label'
 import {useDebounce} from '@skillrecordings/react'
