+++ apps/testingaccessibility/src/components/product-tiers.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import cx from 'classnames'
 import {Pricing} from './pricing'
 import {useCoupon} from '../hooks/use-coupon'
-import type {CommerceProps} from '@skillrecordings/commerce-server'
+import type {CommerceProps} from '@skillrecordings/commerce-server/dist/@types'
 import {PriceCheckProvider} from '../context/pricing-check-context'
 
 export const PricingTiers: React.FC<React.PropsWithChildren<CommerceProps>> = ({
