+++ apps/testingaccessibility/src/context/pricing-check-context.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {type FormattedPrice} from '@skillrecordings/commerce-server'
+import type {FormattedPrice} from '@skillrecordings/commerce-server/dist/@types'
 
 type PricingContextType = {
   addPrice: (price: FormattedPrice, productId: string) => void
