+++ packages/commerce-server/tsup.config.ts
@@ -5,7 +5,7 @@ const isProduction = process.env.NODE_ENV === 'production'
 export default defineConfig({
   clean: true,
   dts: true,
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/@types/index.ts'],
   format: ['cjs', 'esm'],
   minify: isProduction,
   sourcemap: true,
